renamed fastq.rs to fastx.rs
