+++ src/utils/fastx.rs
